+++ _misc/bce31.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python3
 """
+- bce3?.py: 2-part version of bce2:
+  - bce31.py: get data from bitcoin into interim json files
+  - bce32.py: export bce31.py results into interim txt files
+
 Export bitcoind data to gziped jsons.
-3rd parties: None
 
 Usage:
 - [-f, -q, -l]: count txs using short bitcoind reply
