+++ bcepy.py
@@ -2,15 +2,14 @@
 """
 Exports bitcoind data to sql-ready plaintext
 3rd parties:
-- python3-configobj
 - python3-kyotocabinet
-- python3-base58
 - [python3-ujson]
 Bundled:
 - python3-bitcoinrpc
+- python3-base58
 """
 
-from bcepy.main2 import main
+from bcepy import main
 
 if __name__ == "__main__":
     main()
